changed pos of some classes and rewrote stat_bonus meth of equipment
Before stat_bonus was a complexety nightmare. Its still not good but
more readable. A lot of checks have been moved to a seperate mthode,
check_relevant_equip. This methode performs checks if the atribute of
the class is relevant to have an equipment applied to.

Next changes to the application logic. Now a simple modifier is built
and the val to be changed is then multiplied by it. This further reduced
redundancy from before.